+++ packages/core/test/test-util/test-services/v2/test-service/TestEntityEndsWithSomethingElse.ts
@@ -32,7 +32,7 @@ export class TestEntityEndsWithSomethingElse extends EntityV2 implements TestEnt
    * Returns an entity builder to construct instances `TestEntityEndsWithSomethingElse`.
    * @returns A builder that constructs instances of entity type `TestEntityEndsWithSomethingElse`.
    */
-  static builder(): EntityBuilderType<TestEntityEndsWithSomethingElse, TestEntityEndsWithSomethingElseTypeForceMandatory> {
+  static builder(): EntityBuilderType<TestEntityEndsWithSomethingElse, TestEntityEndsWithSomethingElseType> {
     return EntityV2.entityBuilder(TestEntityEndsWithSomethingElse);
   }
 
@@ -63,10 +63,6 @@ export class TestEntityEndsWithSomethingElse extends EntityV2 implements TestEnt
 }
 
 export interface TestEntityEndsWithSomethingElseType {
-  keyProperty: string;
-}
-
-export interface TestEntityEndsWithSomethingElseTypeForceMandatory {
   keyProperty: string;
 }