+++ packages/core/test/test-util/test-services/v2/test-service/TestEntitySingleLink.ts
@@ -62,7 +62,7 @@ export class TestEntitySingleLink extends EntityV2 implements TestEntitySingleLi
    * Returns an entity builder to construct instances `TestEntitySingleLink`.
    * @returns A builder that constructs instances of entity type `TestEntitySingleLink`.
    */
-  static builder(): EntityBuilderType<TestEntitySingleLink, TestEntitySingleLinkTypeForceMandatory> {
+  static builder(): EntityBuilderType<TestEntitySingleLink, TestEntitySingleLinkType> {
     return EntityV2.entityBuilder(TestEntitySingleLink);
   }
 
@@ -101,16 +101,6 @@ export interface TestEntitySingleLinkType {
   booleanProperty?: boolean;
   guidProperty?: string;
   int16Property?: number;
-  toMultiLink: TestEntityLvl2MultiLinkType[];
-  toSingleLink: TestEntityLvl2SingleLinkType;
-}
-
-export interface TestEntitySingleLinkTypeForceMandatory {
-  keyProperty: string;
-  stringProperty: string;
-  booleanProperty: boolean;
-  guidProperty: string;
-  int16Property: number;
   toMultiLink: TestEntityLvl2MultiLinkType[];
   toSingleLink: TestEntityLvl2SingleLinkType;
 }