+++ packages/generator/src/entity/file.ts
@@ -4,10 +4,7 @@ import { SourceFileStructure, StructureKind } from 'ts-morph';
 import { VdmEntity, VdmServiceMetadata } from '../vdm-types';
 import { entityClass } from './class';
 import { importDeclarations, otherEntityImports } from './imports';
-import {
-  entityTypeForceMandatoryInterface,
-  entityTypeInterface
-} from './interface';
+import { entityTypeInterface } from './interface';
 import { entityNamespace } from './namespace';
 
 export function entitySourceFile(