+++ test-packages/test-services/v2/test-service/TestEntityLvl2SingleLink.d.ts
@@ -47,7 +47,7 @@ export declare class TestEntityLvl2SingleLink extends EntityV2 implements TestEn
      * Returns an entity builder to construct instances `TestEntityLvl2SingleLink`.
      * @returns A builder that constructs instances of entity type `TestEntityLvl2SingleLink`.
      */
-    static builder(): EntityBuilderType<TestEntityLvl2SingleLink, TestEntityLvl2SingleLinkTypeForceMandatory>;
+    static builder(): EntityBuilderType<TestEntityLvl2SingleLink, TestEntityLvl2SingleLinkType>;
     /**
      * Returns a request builder to construct requests for operations on the `TestEntityLvl2SingleLink` entity type.
      * @returns A `TestEntityLvl2SingleLink` request builder.
@@ -73,13 +73,6 @@ export interface TestEntityLvl2SingleLinkType {
     booleanProperty?: boolean;
     guidProperty?: string;
     int16Property?: number;
-}
-export interface TestEntityLvl2SingleLinkTypeForceMandatory {
-    keyProperty: string;
-    stringProperty: string;
-    booleanProperty: boolean;
-    guidProperty: string;
-    int16Property: number;
 }
 export declare namespace TestEntityLvl2SingleLink {
     /**