+++ test-packages/test-services/v2/test-service/TestEntityMultiLink.ts
@@ -62,7 +62,7 @@ export class TestEntityMultiLink extends EntityV2 implements TestEntityMultiLink
    * Returns an entity builder to construct instances `TestEntityMultiLink`.
    * @returns A builder that constructs instances of entity type `TestEntityMultiLink`.
    */
-  static builder(): EntityBuilderType<TestEntityMultiLink, TestEntityMultiLinkTypeForceMandatory> {
+  static builder(): EntityBuilderType<TestEntityMultiLink, TestEntityMultiLinkType> {
     return EntityV2.entityBuilder(TestEntityMultiLink);
   }
 
@@ -101,16 +101,6 @@ export interface TestEntityMultiLinkType {
   booleanProperty?: boolean;
   guidProperty?: string;
   int16Property?: number;
-  toMultiLink: TestEntityLvl2MultiLinkType[];
-  toSingleLink: TestEntityLvl2SingleLinkType;
-}
-
-export interface TestEntityMultiLinkTypeForceMandatory {
-  keyProperty: string;
-  stringProperty: string;
-  booleanProperty: boolean;
-  guidProperty: string;
-  int16Property: number;
   toMultiLink: TestEntityLvl2MultiLinkType[];
   toSingleLink: TestEntityLvl2SingleLinkType;
 }