+++ test-packages/test-services/v2/test-service/TestEntityOtherMultiLink.ts
@@ -32,7 +32,7 @@ export class TestEntityOtherMultiLink extends EntityV2 implements TestEntityOthe
    * Returns an entity builder to construct instances `TestEntityOtherMultiLink`.
    * @returns A builder that constructs instances of entity type `TestEntityOtherMultiLink`.
    */
-  static builder(): EntityBuilderType<TestEntityOtherMultiLink, TestEntityOtherMultiLinkTypeForceMandatory> {
+  static builder(): EntityBuilderType<TestEntityOtherMultiLink, TestEntityOtherMultiLinkType> {
     return EntityV2.entityBuilder(TestEntityOtherMultiLink);
   }
 
@@ -63,10 +63,6 @@ export class TestEntityOtherMultiLink extends EntityV2 implements TestEntityOthe
 }
 
 export interface TestEntityOtherMultiLinkType {
-  keyProperty: string;
-}
-
-export interface TestEntityOtherMultiLinkTypeForceMandatory {
   keyProperty: string;
 }